[TASK-21] Refactor question component
