+++ src/screens/SurveyDetail/question.js
@@ -3,9 +3,9 @@ import React, { useContext, useState } from 'react';
 import { DetailsContext } from 'contexts/details';
 import closeIcon from 'images/close-button-white.svg';
 import nextIcon from 'images/next-button-black.svg';
-import DetermineQuestionType from 'screens/SurveyDetail/determineQuestionType';
-import filterQuestionList from 'screens/SurveyDetail/filterQuestionList';
-import questionProperties from 'screens/SurveyDetail/questionProperties';
+import DetermineQuestionType from 'screens/SurveyDetail/QuestionBuilders/determineQuestionType';
+import filterQuestionList from 'screens/SurveyDetail/QuestionBuilders/filterQuestionList';
+import questionProperties from 'screens/SurveyDetail/QuestionBuilders/questionProperties';
 
 const Questions = () => {
   const detailsContext = useContext(DetailsContext);
